+++ CHANGES.md
@@ -1,3 +1,8 @@
+master (in development)
+=======================
+
+- Fix `IOError` if `$XDG_CURRENT_DESKTOP` is missing
+
 0.1.0.0 (Apr 1, 2014)
 =====================